pyproject.toml
[project]
name = "mansour-strategy-evaluation-service"
version = "0.1.0"
description = ""
authors = [
    {name = "doif.dobby",email = "doif.dobby@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "py-eureka-client (>=0.13.0,<0.14.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "redis (>=6.4.0,<7.0.0)",
    "ta-lib (>=0.6.7,<0.7.0)",
    "pandas (>=2.3.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "mansour_strategy_evaluation_service", from = "src"}]

[tool.poetry.scripts]
dev = "mansour_strategy_evaluation_service.main:start"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
